+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Added standardDeviation to NetMath
 - Added gaussian weights distribution
 - Added xavierNormal weights distribution
+- Added lecunUniform weights distribution
   
 # v1.3.0 - Regularization
 ---